+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+
+## [1.1.0] 2018-02-07
+### Added
+- Add default browser list settings.
+- Add exportable basic babel settings.
+- Add rewriting to compile link to src folder to dist folder. (eg. ui-constants/src/ad => ui-constants/dist/ad).
+
+
 ## [1.0.0] 2018-02-05
 Project start